+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Improvements
   DataViewerFrame.
 - Improved the design of the DataBrowsingFrame and selectors for raw or hdf5 
   data are now located directly above the plots.
+- Added type-checking for Datasets axis_ranges and axis_units.  
 
 Bugfixes
 --------